spl/cred: move remaining idmap compat, use generic checks
Move the remaining idmap/cred checks from sys/types.h to sys/cred.h, the
only direct user, and change them to use the generic macros.

These should be folded back into idmap_compat at some point but they
need careful auditing first.

Sponsored-by: TrueNAS
