+++ include/os/linux/spl/sys/types.h
@@ -57,28 +57,6 @@ typedef ulong_t			pgcnt_t;
 typedef int			major_t;
 typedef int			minor_t;
 
-struct user_namespace;
-#ifdef HAVE_IOPS_CREATE_IDMAP
-#include <linux/refcount.h>
-#ifdef HAVE_IDMAP_NO_USERNS
-#include <linux/user_namespace.h>
-struct mnt_idmap {
-	struct uid_gid_map uid_map;
-	struct uid_gid_map gid_map;
-	refcount_t count;
-};
-#define	idmap_owner(p)	(NULL)
-#else
-struct mnt_idmap {
-	struct user_namespace *owner;
-	refcount_t count;
-};
-#define	idmap_owner(p)	(((struct mnt_idmap *)p)->owner)
-#endif
-#else
-#define	idmap_owner(p)	((struct user_namespace *)p)
-#endif
-
 #ifdef HAVE_1ARG_ASSIGN_STR
 #define	__assign_str_impl(a, b)		__assign_str(a)
 #else
